AddNodeMenu Basic Functionality
Major cleanup of the pre-rework script, comments have been added to most pieces of code to elaborate on what is being done.
Certain parts of code were re-written to improve efficiency and readability.

Compared to the previous version this menu requires confirmation to add a node, instead of applying it on selection.
Double clicking is broken due to (assumed) 3DUI limitations, and will be custom-made in another commit.