Add Custom GDScript CodeHighlighter
Still not fully happy with this one, I can't seem to get annotations/similar to work due to color regions having no way to end from various symbols or even just empty space.
I don't wanna have to color everything using a custom function, as I'd have to run that for every text change, but at this point that might be the only way to get the rest of these working.