Giant Mess, Read Description
I have been working on & off on this for quite a while now, finally somewhat settling on some way of doing this.
All of the commits that get pushed with this one are part of this gigantic mess of constant retries on how to actually realize this thing.
There will be a few more somewhat major changes to the way the UI is done, but the basics are mostly in place with no more issues.